修复按id删除的bug
为id字段添加了一个@JsonSerialize(using = ToStringSerializer.class)

+++ src/main/java/com/zlyyl/controller/BookController.java
@@ -31,7 +31,7 @@ public class BookController {
     }
 
     @DeleteMapping("/{id}")
-    public Result delete(@PathVariable Integer id) {
+    public Result delete(@PathVariable Long id) {
         Boolean flag = bookService.delete(id);
         return new Result(flag ? Code.DELETE_OK : Code.DELETE_ERR, flag);
     }
